+++ src/components/forms/FooterNewInvoice.js
@@ -27,7 +27,7 @@ export const FooterNewInvoice = () => {
     }
 
     return (
-        <footer className="flex items-center justify-between rounded-xl bg-white dark:bg-dark-300  md:fixed md:w-[72.8%] md:pr-[56px] md:bottom-0 md:pb-4 lg:static lg:w-full lg:pr-0 lg:mt-0 lg:mb-3">
+        <footer className="flex items-center justify-between rounded-xl bg-white dark:bg-dark-300  fixed md:w-[72.8%] md:pr-[56px] bottom-0 pb-4 lg:static lg:w-full lg:pr-0 lg:mt-0 lg:mb-3">
             <Button styles="bg-neutral-100 text-neutral-500" type="button" title="Discard" onClick={discardHandler} />
             <Button styles="bg-dark-500  text-neutral-300 dark:bg-dark-200 dark:text-neutral-200 md:hidden" type="button" title="Save as Draft" onClick={saveAsDraftHandler} />
             <Button styles="bg-primary-200 text-white md:hidden" type="submit" title="Save & Send" onClick={() => setFieldValue('status', 'pending')} />
